Tweak download URL to point to latest GitHub release
Rather than point the download link to the version of the
actual workflow on the master branch, point to the latest release.

A bit more overhead, but makes a bit more sense, and may allow us
to stop tracking the workflow if we want, since we can just include it
in the GH releases.