Update Makefile to better handle info.plist
Instead of treating the info.plist as a file to be copied around, this
now links it, and the icon-cache directory, into the
`$ALFRED_PKGMAN_WORKFLOW_DIR` so changes to the Workflow inside Alfred
will be reflected in the repo’s copy of `info.plist`.

This means any changes made inside Alfred will end up landing in this
repo, and then can be committed & used to build a release.